[#72] Moved example code to own repository
