--- examples/.gitignore
@@ -1,4 +1,0 @@
-/tmp
-/contents
-/core
-/examples.json
