--- examples/uninstall.sh
@@ -1,5 +1,0 @@
-base=$(dirname $0)
-rm -rf "$base/tmp"
-rm -rf "$base/core"
-rm -rf "$base/contents"
-rm -rf "$base/examples.json"
