+++ package.json
@@ -7,19 +7,14 @@
         "url": "https://github.com/Lumieducation/h5p-nodejs-library"
     },
     "scripts": {
-        "prepare": "examples/install.sh",
-        "start": "node ./examples/server",
-        "clean": "rm -rf build && rm -rf dist",
-        "uninstall": "npm run clean && rm -rf node_modules && examples/uninstall.sh",
+        "clean": "rm -rf build",
+        "uninstall": "npm run clean && rm -rf node_modules",
         "build": "npx babel src --presets babel-preset-env --out-dir build",
         "build:watch": "npx babel src --watch --presets babel-preset-env --out-dir build",
-        "ci": "npm run build && npm run lint && npm run format:check && npm run test && npm run download:content && npm run test:content && npm run test:content:async",
+        "ci": "npm run build && npm run lint && npm run format:check && npm run test",
         "lint": "eslint ./src ./test",
         "test": "jest",
-        "download:content": "node test/download_content.ci.js ",
         "test:watch": "jest --watch",
-        "test:content": "node test/content.ci.js",
-        "test:content:async": "node test/content-with-async-libraryLoader.ci.js",
         "format:check": "prettier --check \"{src,test}/**/*.js\"",
         "format": "prettier --write \"{src,test}/**/*.js\""
     },
@@ -36,13 +31,9 @@
         "eslint-config-prettier": "^4.2.0",
         "eslint-plugin-import": "^2.17.2",
         "eslint-plugin-jest": "^22.6.4",
-        "express": "^4.17.0",
         "jest": "^24.7.1",
-        "jsdom": "^15.1.0",
         "prettier": "^1.17.1",
-        "promise-queue": "^2.2.5",
-        "puppeteer": "^1.19.0",
-        "request-promise": "^4.2.4"
+        "puppeteer": "^1.19.0"
     },
     "main": "./src/index.js"
 }
